Fix instances of -Wshift-negative-value
No changes in code generation

+++ iop/memorycard/mcman/src/main.c
@@ -3435,7 +3435,7 @@ lbl0:
 			return sceMcResSucceed;
 		}
 
-		if ((mce->wr_flag & ~(-1 << j)) == 0) {
+		if ((mce->wr_flag & ~(((u32)(-1)) << j)) == 0) {
 			mce->wr_flag = 0;
 			return sceMcResSucceed;
 		}
@@ -4054,7 +4054,7 @@ int McReplaceBadBlock(void)
 		}
 
 		i = 0; //sp5c
-		value = ~(-1 << mcdi->clusters_per_block);
+		value = ~(((u32)(-1)) << mcdi->clusters_per_block);
 		do {
 			clust = (mcman_badblock * mcdi->clusters_per_block) + i;
 			if ((u32)clust < mcdi->alloc_offset) {
